+++ src/app/(provider)/(root)/(community)/layout.tsx
@@ -5,7 +5,7 @@ interface LayoutProps {
 }
 
 export default function CommunityLayout({ children }: LayoutProps) {
-  return <>{children}</>;
+  return <div className="w-[1200px] mx-auto flex">{children}</div>;
 }
 
 // 일단 대충 mx auto 먹여서 가운데 정렬 하려고 했더니 안되네 왜지 딴사람들거 봐야겠다
