+++ src/app/(provider)/(root)/(community)/qna/[id]/page.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import CommuCommentList from '../../_components/CommuCommentList';
 import CommuComment from '../../_components/CommuComment';
 import CommuPost from '../../_components/CommuPost';
 
@@ -11,6 +12,7 @@ export default function page() {
       <div className="flex flex-col gap-16">
         <CommuPost />
         <CommuComment />
+        <CommuCommentList />
       </div>
     </div>
   );
