+++ src/app/api/community/route.ts
@@ -1,0 +1,12 @@
+import { createClient } from '@/utils/supabase/server';
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  try {
+    const supabase = createClient();
+    const { data } = await supabase.from('Community Posts').select('*');
+    return NextResponse.json(data);
+  } catch (error) {
+    return NextResponse.json({ error: '데이터를 가져오는 데 실패했습니다.' });
+  }
+}
